requirements.txt
absl-py==2.1.0
aioredis==2.0.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
astor==0.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrdict==2.0.1
Babel==2.15.0
bce-python-sdk==0.9.11
beautifulsoup4==4.12.3
billiard==4.2.0
blinker==1.8.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.2.1
cssselect==1.2.0
cssutils==2.11.0
cycler==0.12.1
Cython==3.0.10
decorator==5.1.1
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpeg-python==0.2.0
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flask-babel==4.0.0
fonttools==4.52.4
fsspec==2024.3.1
future==1.0.0
futures==3.0.5
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
kiwisolver==1.4.5
kombu==5.3.7
lazy_loader==0.4
lmdb==1.4.1
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
moviepy==1.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.1.105
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
orjson==3.10.3
packaging==24.0
paddleocr==2.7.3
paddlepaddle-gpu==2.6.1
pandas==2.2.2
parso==0.8.4
pdf2docx==0.5.8
pexpect==4.9.0
pillow==10.3.0
premailer==3.10.0
proglog==0.1.10
prompt-toolkit==3.0.43
protobuf==5.27.0
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyclipper==1.3.0.post5
pycryptodome==3.20.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyMuPDF==1.24.4
PyMuPDFb==1.24.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.1
rarfile==4.2
redis==5.0.4
requests==2.32.2
rich==13.7.1
scikit-image==0.22.0
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
thop==0.1.1.post2209072238
tifffile==2024.5.22
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.4
traitlets==5.14.2
triton==2.2.0
typer==0.12.3
typing_extensions==4.12.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.0
uvloop==0.19.0
vine==5.1.0
visualdl==2.5.3
watchdog==5.0.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
zipp==3.19.0
